+++ CHANGELOG.md
@@ -43,6 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix unusable version.dist variable in packaging code templates (#116).
 - Fix crash on build on unsupported derivative in instance pipelines (#117).
 - Fix crash on retrieving derivative version in artifact definition (#118).
+- Fix crash when main archive source is not defined on build of Deb or RPM
+  package (#119).
 
 ## [2.0.0] - 2023-05-05